[UPDATE] Like

+++ src/main/java/com/example/sinabroproject/entity/like/LikeRepository.java
@@ -2,5 +2,5 @@ package com.example.sinabroproject.entity.like;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LikeRepository extends JpaRepository<Like,Long> {
+public interface LikeRepository extends JpaRepository<Like,Integer> {
 }
